requirements-dev.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
coverage==7.13.5
defusedxml==0.7.1
fastapi==0.109.0
h11==0.16.0
httpcore==1.0.9
httpie==3.2.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonschema==4.21.1
jsonschema-specifications==2025.9.1
kafka-python-ng==2.2.3
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.20.0
PySocks==1.7.1
pytest==8.0.0
pytest-cov==4.1.0
PyYAML==6.0.1
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
starlette==0.35.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
